custom-elements/custom-element-lifecycle/types-of-callbacks/attached-callback-move-element-test.html

<!DOCTYPE html>
<html>
<head>
<title>Attached callback of a custom element should be called if element is moved</title>
<meta name="author" title="Aleksei Yu. Semenov" href="mailto:a.semenov@unipro.ru">
<meta name="assert" content="attached callback must be enqueued whenever custom element is inserted into a document and this document has a browsing context">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#types-of-callbacks">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">
test(function() {
    var doc = newHTMLDocument();
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.attachedCallback = function() {
        callbackCalledCounter++;
    }
    doc.registerElement('x-a-1', {prototype: proto});
    var customElement = doc.createElement('x-a-1');
    doc.body.appendChild(customElement);
    assert_equals(callbackCalledCounter, 0, 'attachedCallback should not be called '+
        'in documents that do not have a browsing context');

    var divElement = doc.createElement('div');
    doc.body.appendChild(divElement);
    divElement.appendChild(customElement);
    assert_equals(callbackCalledCounter, 0, 'attachedCallback should not be called '+
        'in documents that do not have a browsing context');
}, 'Test attached callback when moving custom element inside document '+
    'without browsing context');


testInIFrame('../../resources/blank.html', function(docBC) {
    var docNoBC = newHTMLDocument();
    var callbackCalledCounter = 0;

    var proto1 = Object.create(HTMLElement.prototype);
    proto1.attachedCallback = function() {
        callbackCalledCounter++;
    }
    docNoBC.registerElement('x-a-2', {prototype: proto1});
    var customElement = docNoBC.createElement('x-a-2');
    docNoBC.body.appendChild(customElement);
    assert_equals(callbackCalledCounter, 0, 'attachedCallback should not be called '+
        'in documents that do not have a browsing context');

    var proto2 = Object.create(HTMLElement.prototype);
    proto2.attachedCallback = function() {
        callbackCalledCounter++;
    }
    docBC.registerElement('x-a-2', {prototype: proto2});
    docBC.body.appendChild(customElement);
    assert_equals(callbackCalledCounter, 1, 'attachedCallback should be called '+
        'in documents with browsing context');
}, 'Test attached callback when moving custom element from '+
    'document without browsing context to document with browsing context');


testInIFrame('../../resources/blank.html', function(docBC) {
    var callbackCalledCounter = 0;

    var proto1 = Object.create(HTMLElement.prototype);
    proto1.attachedCallback = function() {
        callbackCalledCounter++;
    }
    docBC.registerElement('x-a-3', {prototype: proto1});
    var customElement = docBC.createElement('x-a-3');
    docBC.body.appendChild(customElement);
    assert_equals(callbackCalledCounter, 1, 'attachedCallback should be called '+
        'in documents with browsing context');

    var docNoBC = newHTMLDocument();
    var proto2 = Object.create(HTMLElement.prototype);
    proto2.attachedCallback = function() {
        callbackCalledCounter++;
    }
    docNoBC.registerElement('x-a-3', {prototype: proto2});
    docNoBC.body.appendChild(customElement);
    assert_equals(callbackCalledCounter, 1, 'attachedCallback should not be called '+
        'in documents that do not have a browsing context');
}, 'Test attached callback when moving custom element from '+
    'document with browsing context to document without browsing context');


testInIFrame('../../resources/blank.html', function(doc) {
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.attachedCallback = function() {
        callbackCalledCounter++;
    }
    doc.registerElement('x-a-4', {prototype: proto});
    var customElement = doc.createElement('x-a-4');
    doc.body.appendChild(customElement);
    assert_equals(callbackCalledCounter, 1, 'attachedCallback should be called '+
        'in documents with browsing context');

    var divElement = doc.createElement('div');
    doc.body.appendChild(divElement);
    divElement.appendChild(customElement);
    assert_equals(callbackCalledCounter, 2, 'attachedCallback should be called '+
        'in documents with browsing context');
}, 'Test attached callback when moving custom element inside document '+
    'with browsing context');


var moveTest1 =  async_test('Test attached callback when moving custom element from '+
    'document with browsing context to document with browsing context');

moveTest1.testFunction = function() {
    // TODO
}

moveTest1.step(function() {
    moveTest1.callbackCalledCounter = 0;
    moveTest1.readyToTest = 0;

    var iframe1 = document.createElement('iframe');
    iframe1.src = '../../resources/blank.html';
    document.body.appendChild(iframe1);
    iframe1.onload = t.step_func(function() {
        var doc1 = iframe1.contentDocument;
        var proto = Object.create(HTMLElement.prototype);
        proto.attachedCallback = function() {
            callbackCalledCounter++;
        }
        doc1.registerElement('x-a-6', {prototype: proto});
        moveTest1.customElement = doc1.createElement('x-a-6');
        doc1.body.appendChild(moveTest1.customElement);
        assert_equals(callbackCalledCounter, 1, 'attachedCallback should be called '+
            'in documents with browsing context');

        moveTest1.readyToTest++;
        if (moveTest1.readyToTest==2) {
            moveTest1.testFunction();
            moveTest1.done();
        }
    });

    var iframe2 = document.createElement('iframe');
    iframe1.src = '../../resources/x-element.html';
    document.body.appendChild(iframe2);
    iframe1.onload = t.step_func(function() {
        var doc1 = iframe1.contentDocument;
        var proto = Object.create(HTMLElement.prototype);
        proto.attachedCallback = function() {
            callbackCalledCounter++;
        }
        doc1.registerElement('x-a-6', {prototype: proto});
        moveTest1.customElement = doc1.createElement('x-a-6');
        doc1.body.appendChild(moveTest1.customElement);
        assert_equals(callbackCalledCounter, 1, 'attachedCallback should be called '+
            'in documents with browsing context');

        moveTest1.readyToTest++;
        if (moveTest1.readyToTest==2) {
            moveTest1.testFunction();
            moveTest1.done();
        }
    });

});


    var proto = Object.create(HTMLElement.prototype);
    proto.attachedCallback = function() {
        callbackCalledCounter++;
    }
    doc1.registerElement('x-a-5', {prototype: proto});
    var customElement = doc1.createElement('x-a-5');
    doc1.body.appendChild(customElement);
    assert_equals(callbackCalledCounter, 1, 'attachedCallback should be called '+
        'in documents with browsing context');





}, );


/*

test(function() {
    var doc = newHTMLDocument();
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.attachedCallback = function() {
        callbackCalledCounter++;
    }
    var GeneratedConstructor = doc.registerElement('x-a-1', {prototype: proto});
    var customElement = new GeneratedConstructor();
    assert_equals(callbackCalledCounter, 0, 'attachedCallback should not be called '+
        'in documents that do not have a browsing context');

}, 'Test attached callback when custom element is instantiated via constructor. ' +
    'Document has no browsing context');


test(function() {
    var doc = newHTMLDocument();
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.attachedCallback = function() {
        callbackCalledCounter++;
    }
    doc.registerElement('x-a-2', {prototype: proto});
    doc.body.innerHTML = '<x-a-2></x-a-2>';
    assert_equals(callbackCalledCounter, 0, 'attachedCallback should not be called '+
        'in documents that do not have a browsing context');

}, 'Test attached callback when custom element is created via innerHTML property. ' +
    'Document has no browsing context');


test(function() {
    var doc = newHTMLDocument();
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.attachedCallback = function() {
        callbackCalledCounter++;
    }
    doc.body.innerHTML = '<x-a-3></x-a-3>';
    doc.registerElement('x-a-3', {prototype: proto});
    assert_equals(callbackCalledCounter, 0, 'attachedCallback should not be called '+
        'in documents that do not have a browsing context');

}, 'Test attached callback when custom element is created via innerHTML property before ' +
    'registration. Document has no browsing context');


test(function() {
    var doc = newHTMLDocument();
    var callbackCalledCounter = 0;

    doc.body.innerHTML = '<x-a-4 id="xa"></x-a-4>';
    var customElement = doc.querySelector('#xa');
    var proto = Object.create(HTMLElement.prototype);
    proto.attachedCallback = function() {
        callbackCalledCounter++;
    }
    customElement.constructor.prototype = proto;
    assert_equals(callbackCalledCounter, 0, 'Callback attached should not be called ' +
        'for unregistered custom element in document without browsing context');

}, 'Test attached callback when custom element is unregistered.');


testInIFrame('../../resources/x-element.html', function(doc) {
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.attachedCallback = function() {
        callbackCalledCounter++;
    }
    doc.registerElement('x-element', {prototype: proto});
    assert_equals(callbackCalledCounter, 1, 'attachedCallback should be called '+
        'in documents with browsing context');

}, 'Test attached callback. Document has browsing context');


testInIFrame('../../resources/blank.html', function(doc) {
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.attachedCallback = function() {
        callbackCalledCounter++;
    }
    doc.registerElement('x-element', {prototype: proto});
    var x = doc.createElement('x-element');
    assert_equals(callbackCalledCounter, 0, 'attachedCallback should not be called '+
        'before element is added to document with browsing context');
    doc.body.appendChild(x);
    assert_equals(callbackCalledCounter, 1, 'attachedCallback should be called '+
        'in documents with browsing context');

}, 'Test attached callback. Registered element is created via document.createElement(). ' +
    'Document has browsing context');


testInIFrame('../../resources/blank.html', function(doc) {
    var callbackCalledCounter = 0;

    var x = doc.createElement('x-element');
    doc.body.appendChild(x);
    var proto = Object.create(HTMLElement.prototype);
    proto.attachedCallback = function() {
        callbackCalledCounter++;
    }
    doc.registerElement('x-element', {prototype: proto});
    assert_equals(callbackCalledCounter, 1, 'attachedCallback should be called '+
        'in documents with browsing context');

}, 'Test attached callback. Unregistered element is created via document.createElement(). ' +
    'Document has browsing context');


testInIFrame('../../resources/blank.html', function(doc) {
    var callbackCalledCounter = 0;

    var proto = Object.create(HTMLElement.prototype);
    proto.attachedCallback = function() {
        callbackCalledCounter++;
    }
    doc.registerElement('x-element', {prototype: proto});
    doc.body.innerHTML = '<x-element></x-element>';
    assert_equals(callbackCalledCounter, 1, 'attachedCallback should be called '+
        'in documents with browsing context');

}, 'Test attached callback. Registered element is created via innerHTML property. ' +
    'Document has browsing context');


testInIFrame('../../resources/blank.html', function(doc) {
    var callbackCalledCounter = 0;

    doc.body.innerHTML = '<x-element></x-element>';
    var proto = Object.create(HTMLElement.prototype);
    proto.attachedCallback = function() {
        callbackCalledCounter++;
    }
    doc.registerElement('x-element', {prototype: proto});
    assert_equals(callbackCalledCounter, 1, 'attachedCallback should be called '+
        'in documents with browsing context');

}, 'Test attached callback. Unresolved element is created via innerHTML property. ' +
    'Document has browsing context');
*/
</script>
</body>
</html>
